+++ contracts/Interfaces.sol
@@ -3,17 +3,18 @@
 pragma solidity ^0.8.9;
 
 struct Bag {
-    uint256 totalManaProduced;
-    uint256 mintCount;
+    uint64 totalManaProduced;
+    uint64 mintCount;
 }
 
 struct Crystal {
     uint16 attunement;
     uint64 lastClaim;
     uint64 lastLevelUp;
-    uint64 level;
-    uint256 levelManaProduced;
-    uint256 regNum;
+    uint8 level;
+    uint32 levelManaProduced;
+    uint16 regNum;
+
 }
 
 struct Collab {
@@ -23,16 +24,16 @@ struct Collab {
 }
 
 interface ICrystalManaCalculator {
-    function claimableMana(uint256 tokenId) external view returns (uint256);
+    function claimableMana(uint256 tokenId) external view returns (uint32);
 }
 
 interface ICrystals {
     function crystalsMap(uint256 tokenID) external view returns (Crystal memory);
     function bags(uint256 tokenID) external view returns (Bag memory);
     function collabMap(uint256 tokenID) external view returns (Collab memory);
-    function getResonance(uint256 tokenId) external view returns (uint256);
-    function getSpin(uint256 tokenId) external view returns (uint256);
-    function claimableMana(uint256 tokenID) external view returns (uint256);
+    function getResonance(uint256 tokenId) external view returns (uint32);
+    function getSpin(uint256 tokenId) external view returns (uint32);
+    function claimableMana(uint256 tokenID) external view returns (uint32);
     function safeTransferFrom(address from, address to, uint256 tokenId, bytes calldata data) external;
     function transferFrom(address from, address to, uint256 tokenId) external;
 }
